# Inplace cosserat stress