--- a/.gitignore
+++ b/.gitignore
@@ -29,4 +29,4 @@ Manifest.toml
 
 # Benchmark files.
 tune.json
-results.md
+*results.md