--- a/src/Materials/HyperElasticMaterial.jl
+++ b/src/Materials/HyperElasticMaterial.jl
@@ -2,12 +2,11 @@
 module HyperElasticMaterial
 
 using Tensors, Reexport
-
 using ..HyperElasticMaterials
 using ..Utils
 
 @reexport import ..Materials: parameters
-@reexport import ..HyperElasticMaterials: cosserat_stress, strain_energy
+@reexport import ..HyperElasticMaterials: cosserat_stress!, strain_energy
 
 export HyperElastic
 
@@ -45,26 +44,24 @@ parameters(m::HyperElastic) = m.params
 "Return the Cosserat or Second-Piola Kirchoff stress tensor `𝕊`
 considering a `SVK` material `m` and the Lagrangian Green
 strain tensor `𝔼`.Also this function provides `∂𝕊∂𝔼` for the iterative method."
-function cosserat_stress(m::HyperElastic, 𝔼::AbstractMatrix)
-    𝔼 = SymmetricTensor{2,3}(𝔼)
-
+function cosserat_stress!(S::AbstractMatrix{<:Real}, ∂S∂E::AbstractMatrix{<:Real},
+                          m::HyperElastic, E::AbstractMatrix)
+    # Transform 𝔼 to a Tenor
+    𝔼 = SymmetricTensor{2,3}(E)
     # Closure strain energy function
     Ψ = E -> strain_energy(m)(E, parameters(m)...)
 
     ∂²Ψ∂E², 𝕊 = hessian(Ψ, 𝔼, :all)
+    # Fill symmetric matrix
+    fill_symmetric_matrix!(S, 𝕊[1, 1], 𝕊[2, 2], 𝕊[3, 3], 𝕊[2, 3], 𝕊[1, 3], 𝕊[1, 2])
 
     # Fill ∂S∂𝔼 with Belischko nomenclature
-    ∂𝕊∂𝔼 = zeros(6, 6)
-    indexes = [(1, 1), (2, 2), (3, 3), (2, 3), (1, 3), (1, 2)]
-
     row = 1
-    for index in indexes
-        i, j = index
-        ∂𝕊∂𝔼[row, :] .= voigt(∂²Ψ∂E²[:, :, i, j])
+    for (i, j) in INDEXES_TO_VOIGT
+        ∂S∂E[row, :] .= voigt(∂²Ψ∂E²[:, :, i, j])
         row += 1
     end
-
-    𝕊, ∂𝕊∂𝔼
+    S, ∂S∂E
 end
 
 end