--- a/src/Materials/HyperElasticMaterials.jl
+++ b/src/Materials/HyperElasticMaterials.jl
@@ -3,7 +3,7 @@ module HyperElasticMaterials
 
 using ..Materials: AbstractMaterial
 
-export AbstractHyperElasticMaterial, cosserat_stress, strain_energy
+export AbstractHyperElasticMaterial, cosserat_stress!, strain_energy
 
 """ Abstract supertype for all hyper-elastic material models.
 
@@ -12,7 +12,7 @@ These materials are characterized by a strain energy function ψ  that depends o
 
 **Abstract Methods**
 * [`strain_energy`](@ref)
-* [`cosserat_stress`](@ref)
+* [`cosserat_stress!`](@ref)
 
 **Abstract fields**
 * label
@@ -26,6 +26,6 @@ function strain_energy(m::AbstractHyperElasticMaterial, 𝔼) end
 
 "Return the Cosserat or Second-Piola Kirchhoff stress tensor `𝕊` given an `AbstractMaterial` `m` and the
 Green-Lagrange strain tensor `𝔼`."
-function cosserat_stress(m::AbstractMaterial, 𝔼::AbstractMatrix) end
+function cosserat_stress!(m::AbstractMaterial, 𝔼::AbstractMatrix) end
 
 end