Closes #437 

### Main
```julia
"uniaxial_extension" => 5-element BenchmarkTools.BenchmarkGroup:
    "solve, ms = 0.3, nelems = 580, nnodes = 201" => Trial(16.441 ms)
    "solve, ms = 0.1, nelems = 9292, nnodes = 2147" => Trial(300.951 ms)
    "solve, ms = 0.2, nelems = 1343, nnodes = 400" => Trial(39.159 ms)
    "solve, ms = 0.5, nelems = 144, nnodes = 62" => Trial(3.747 ms)
    "solve, ms = 0.4, nelems = 256, nnodes = 107" => Trial(7.018 ms)
"uniaxial_compression" => 5-element BenchmarkTools.BenchmarkGroup:
    "solve, ms = 0.3, nelems = 580, nnodes = 201" => Trial(14.801 ms)
    "solve, ms = 0.1, nelems = 9292, nnodes = 2147" => Trial(281.064 ms)
    "solve, ms = 0.2, nelems = 1343, nnodes = 400" => Trial(35.352 ms)
    "solve, ms = 0.5, nelems = 144, nnodes = 62" => Trial(3.412 ms)
    "solve, ms = 0.4, nelems = 256, nnodes = 107" => Trial(6.384 ms)
```

### This branch
```julia
"uniaxial_extension" => 5-element BenchmarkTools.BenchmarkGroup:
    tags: []
    "solve, ms = 0.3, nelems = 580, nnodes = 201" => Trial(16.355 ms)
    "solve, ms = 0.1, nelems = 9292, nnodes = 2147" => Trial(294.047 ms)
    "solve, ms = 0.2, nelems = 1343, nnodes = 400" => Trial(38.229 ms)
    "solve, ms = 0.5, nelems = 144, nnodes = 62" => Trial(3.793 ms)
    "solve, ms = 0.4, nelems = 256, nnodes = 107" => Trial(7.040 ms)
"uniaxial_compression" => 5-element BenchmarkTools.BenchmarkGroup:
    tags: []
    "solve, ms = 0.3, nelems = 580, nnodes = 201" => Trial(14.415 ms)
    "solve, ms = 0.1, nelems = 9292, nnodes = 2147" => Trial(275.075 ms)
    "solve, ms = 0.2, nelems = 1343, nnodes = 400" => Trial(34.459 ms)
    "solve, ms = 0.5, nelems = 144, nnodes = 62" => Trial(3.317 ms)
    "solve, ms = 0.4, nelems = 256, nnodes = 107" => Trial(6.192 ms)
```