@mvanzulli 2023-08-12
Main branch:
| ID                                                                                                                     | time            | GC time   | memory          | allocations |
|------------------------------------------------------------------------------------------------------------------------|----------------:|----------:|----------------:|------------:|
| `["linear_cylinder_internal_pressure", "solve!, ms = 0.5, nelems = 2365, nnodes = 823"]`                               |  79.648 ms (5%) |  9.829 ms | 176.94 MiB (1%) |      713174 |
| `["linear_cylinder_internal_pressure", "solve!, ms = 1.0, nelems = 10708, nnodes = 3082"]`                             |    1.223 s (5%) | 43.827 ms | 800.02 MiB (1%) |     3217971 |
| `["linear_extension", "solve, ms = 0.1, nelems = 9292, nnodes = 2147"]`                                                | 341.231 ms (5%) | 33.139 ms | 690.18 MiB (1%) |     2759548 |
| `["linear_extension", "solve, ms = 0.2, nelems = 1343, nnodes = 400"]`                                                 |  42.869 ms (5%) |  4.829 ms |  99.85 MiB (1%) |      400138 |
| `["linear_extension", "solve, ms = 0.3, nelems = 580, nnodes = 201"]`                                                  |  24.192 ms (5%) |           |  43.29 MiB (1%) |      173154 |
| `["linear_extension", "solve, ms = 0.4, nelems = 256, nnodes = 107"]`                                                  |   7.642 ms (5%) |           |  19.23 MiB (1%) |       76748 |
| `["linear_extension", "solve, ms = 0.5, nelems = 144, nnodes = 62"]`                                                   |   4.271 ms (5%) |           |  10.81 MiB (1%) |       43252 |
| `["uniaxial_compression", "solve, ms = 0.1, nelems = 9292, nnodes = 2147"]`                                            | 275.075 ms (5%) | 29.486 ms | 547.91 MiB (1%) |     1858794 |
| `["uniaxial_compression", "solve, ms = 0.2, nelems = 1343, nnodes = 400"]`                                             |  34.459 ms (5%) |           |  79.20 MiB (1%) |      270018 |
| `["uniaxial_compression", "solve, ms = 0.3, nelems = 580, nnodes = 201"]`                                              |  14.415 ms (5%) |           |  34.31 MiB (1%) |      116991 |
| `["uniaxial_compression", "solve, ms = 0.4, nelems = 256, nnodes = 107"]`                                              |   6.192 ms (5%) |           |  15.27 MiB (1%) |       51965 |
| `["uniaxial_compression", "solve, ms = 0.5, nelems = 144, nnodes = 62"]`                                               |   3.317 ms (5%) |           |   8.56 MiB (1%) |       29299 |
| `["uniaxial_extension", "solve, ms = 0.1, nelems = 9292, nnodes = 2147"]`                                              | 294.047 ms (5%) | 20.930 ms | 385.14 MiB (1%) |     1654370 |
| `["uniaxial_extension", "solve, ms = 0.2, nelems = 1343, nnodes = 400"]`                                               |  38.229 ms (5%) |           |  55.67 MiB (1%) |      240472 |
| `["uniaxial_extension", "solve, ms = 0.3, nelems = 580, nnodes = 201"]`                                                |  16.355 ms (5%) |           |  24.15 MiB (1%) |      104231 |
| `["uniaxial_extension", "solve, ms = 0.4, nelems = 256, nnodes = 107"]`                                                |   7.040 ms (5%) |           |  10.78 MiB (1%) |       46333 |
| `["uniaxial_extension", "solve, ms = 0.5, nelems = 144, nnodes = 62"]`                                                 |   3.793 ms (5%) |           |   6.04 MiB (1%) |       26131 |