@mforets 2023-08-19
fine with the changes, but the result is not noticeable on such bench, right? 

did you compare the allocs as well?