@mvanzulli 2023-08-21
> fine with the changes, but the result is not noticeable on such bench, right?
> 
> did you compare the allocs as well?

Yes, however there are not noticeable differences in allocations. Tenosors.jl does not provide an in-place operation computing the derivatives via Automatic Differentiation